requirements.txt
dearpygui>=1.10.1
cryptography>=41.0.0
python-dotenv>=1.0.0
